Fixed timing bug in communication between record and device support. In previous releases device support was setting the .ACQG field to 0 (Done) before the final read of the device was complete.  This meant that channel access clients that were doing cagets() of .ACQG to see when acquisition was complete and then reading data immediately could get incorrect data.  Note that this problem did not occur if clients were using monitors (callbacks), because the record did not call dbPostEvents until the final read of data from the device was complete.
The fix to this problem required re-writing the interface between the record and
device support so that device support does not directly modify the .ACQG field.
Rather it now passes a structure with status information which the record uses
to modify the record fields, and it does not set the .ACQG field until the record contains
the latest data.

Changed order of calls to dbPostEvents so ACQG is posted last.